Initial reorg of test harness into separate package to split things up - not working yet
