+++ .gitignore
@@ -32,3 +32,7 @@ nbproject/project.properties
 /andhow-test-harness/target/
 /andhow-system-test-harness/nbproject/
 /andhow-system-test-harness/target/
+/andhow-test-harness copy/nbproject/
+/andhow-test-harness copy/target/
+/andhow-system-tests/nbproject/
+/andhow-system-tests/target/
