--- andhow/src/test/java/org/yarnandtail/andhow/example/restclient/SampleRestClientAppTest.java
@@ -1,137 +1,0 @@
-package org.yarnandtail.andhow.example.restclient;
-
-import java.util.ArrayList;
-import static org.junit.Assert.*;
-import org.junit.Before;
-import org.junit.Test;
-import org.springframework.mock.jndi.SimpleNamingContextBuilder;
-import org.yarnandtail.andhow.AndHow;
-import org.yarnandtail.andhow.AndHowTestBase;
-import static org.yarnandtail.andhow.AndHowTestBase.reloader;
-import org.yarnandtail.andhow.api.AppFatalException;
-import org.yarnandtail.andhow.api.Property;
-import org.yarnandtail.andhow.internal.LoaderProblem;
-import org.yarnandtail.andhow.internal.ValueProblem;
-import org.yarnandtail.andhow.load.JndiLoader;
-import org.yarnandtail.andhow.load.PropertyFileOnClasspathLoader;
-import org.yarnandtail.andhow.load.StringArgumentLoader;
-
-/**
- *
- * @author eeverman
- */
-public class SampleRestClientAppTest extends AndHowTestBase {
-	
-	String[] cmdLineArgs = new String[0];
-	private static final String GROUP_PATH = "org.yarnandtail.andhow.example.restclient.SampleRestClientGroup";
-
-	
-	@Before
-	public void setup() {
-		
-		cmdLineArgs = new String[0];
-		
-	}
-	
-	@Test
-	public void testAllValuesAreSet() {
-		
-		cmdLineArgs = new String[] {
-			GROUP_PATH + ".CLASSPATH_PROP_FILE" + StringArgumentLoader.KVP_DELIMITER + 
-				"/org/yarnandtail/andhow/example/restclient/all.points.speced.properties"
-		};
-				
-				
-		AndHow.builder()
-				.group(SampleRestClientGroup.class)
-				.addCmdLineArgs(cmdLineArgs)
-				.loader(new PropertyFileOnClasspathLoader(SampleRestClientGroup.CLASSPATH_PROP_FILE))
-				.reloadForNonPropduction(reloader);
-		
-		assertEquals("/org/yarnandtail/andhow/example/restclient/all.points.speced.properties", 
-				SampleRestClientGroup.CLASSPATH_PROP_FILE.getValue());
-		assertEquals("  Big App  ", SampleRestClientGroup.APP_NAME.getValue());
-		assertEquals("aquarius.usgs.gov", SampleRestClientGroup.REST_HOST.getValue());
-		assertEquals(new Integer(8080), SampleRestClientGroup.REST_PORT.getValue());
-		assertEquals("doquery/", SampleRestClientGroup.REST_SERVICE_NAME.getValue());
-		assertEquals("abc123", SampleRestClientGroup.AUTH_KEY.getValue());
-		assertEquals(new Integer(4), SampleRestClientGroup.RETRY_COUNT.getValue());
-		assertFalse(SampleRestClientGroup.REQUEST_META_DATA.getValue());
-		assertTrue(SampleRestClientGroup.REQUEST_SUMMARY_DATA.getValue());
-
-	}
-	
-	@Test
-	public void testMinimumPropsAreSet() {
-		
-		cmdLineArgs = new String[] {
-			GROUP_PATH + ".CLASSPATH_PROP_FILE" + StringArgumentLoader.KVP_DELIMITER + 
-				"/org/yarnandtail/andhow/example/restclient/minimum.points.speced.properties"
-		};
-				
-				
-		AndHow.builder()
-				.group(SampleRestClientGroup.class)
-				.addCmdLineArgs(cmdLineArgs)
-				.loader(new PropertyFileOnClasspathLoader(SampleRestClientGroup.CLASSPATH_PROP_FILE))
-				.reloadForNonPropduction(reloader);
-		
-		assertEquals("/org/yarnandtail/andhow/example/restclient/minimum.points.speced.properties", 
-				SampleRestClientGroup.CLASSPATH_PROP_FILE.getValue());
-		assertEquals("aquarius.usgs.gov", SampleRestClientGroup.REST_HOST.getValue());
-		assertEquals(new Integer(8080), SampleRestClientGroup.REST_PORT.getValue());
-		assertEquals("query/", SampleRestClientGroup.REST_SERVICE_NAME.getValue());	//a default value
-		assertEquals("abc123", SampleRestClientGroup.AUTH_KEY.getValue());
-		assertEquals(new Integer(2), SampleRestClientGroup.RETRY_COUNT.getValue());	//a default
-		assertTrue(SampleRestClientGroup.REQUEST_META_DATA.getValue());	//a default
-		assertFalse(SampleRestClientGroup.REQUEST_SUMMARY_DATA.getValue());	//a default
-
-	}
-	
-	@Test
-	public void testInvalidValues() throws Exception {
-		
-		SimpleNamingContextBuilder jndi = AndHowTestBase.getJndi();
-		jndi.activate();
-		
-		cmdLineArgs = new String[] {
-			GROUP_PATH + ".CLASSPATH_PROP_FILE" + StringArgumentLoader.KVP_DELIMITER + 
-				"/org/yarnandtail/andhow/example/restclient/invalid.properties"
-		};
-				
-				
-		try {
-			
-			//Error expected b/c some values are invalid
-			AndHow.builder()
-					.group(SampleRestClientGroup.class)
-					.addCmdLineArgs(cmdLineArgs)
-					.loader(new PropertyFileOnClasspathLoader(SampleRestClientGroup.CLASSPATH_PROP_FILE))
-					.loader(new JndiLoader())
-					.reloadForNonPropduction(reloader);
-		} catch (AppFatalException e) {
-			
-			//Value Problems (validation)
-			//Due to loading from a prop file, the order of the file is not preserved,
-			//so we cannot know the order that problems were encountered.
-			ArrayList<Property<?>> expectedProblemPoints = new ArrayList();
-			expectedProblemPoints.add(SampleRestClientGroup.REST_HOST);
-			expectedProblemPoints.add(SampleRestClientGroup.REST_PORT);
-			expectedProblemPoints.add(SampleRestClientGroup.REST_SERVICE_NAME);
-			
-			assertEquals(3, e.getProblems().filter(ValueProblem.class).size());
-			assertTrue(expectedProblemPoints.contains(e.getProblems().filter(ValueProblem.class).get(0).getBadValueCoord().getProperty()));
-			assertTrue(expectedProblemPoints.contains(e.getProblems().filter(ValueProblem.class).get(1).getBadValueCoord().getProperty()));
-			assertTrue(expectedProblemPoints.contains(e.getProblems().filter(ValueProblem.class).get(2).getBadValueCoord().getProperty()));
-			
-			//
-			// Loader problems
-			assertEquals(1, e.getProblems().filter(LoaderProblem.class).size());
-			assertEquals(SampleRestClientGroup.RETRY_COUNT, e.getProblems().filter(LoaderProblem.class).get(0).getBadValueCoord().getProperty());
-		}
-		
-
-	}
-	
-
-}
